Refactor car components: remove CarCard and CarCardSkeleton, introduce new CarCard and CarCardListingSkeleton; update CarForm to exclude description; adjust home and listings pages for new structure; clean up unused code and improve loading states.
