+++ src/components/car-posting-section.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React from 'react';
-import CarCard from './cards/card-card.tsx';
-import CarCardListingSkeleton from './cards/car-card -listing-skeleton.tsx';
+import CarCard from './cards/card.tsx';
+import CarCardListingSkeleton from './cards/card-listing-skeleton.tsx';
 
 interface CarPostingSectionProps {
     title: string;
